# nodejs-server/api/openapi.yaml
openapi: 3.0.1
info:
  title: Estuary API
  description: This is the API for the Estuary application.
  termsOfService: http://estuary.tech
  contact:
    name: API Support
    url: https://docs.estuary.tech/feedback
  license:
    name: Apache 2.0 Apache-2.0 OR MIT
    url: https://github.com/application-research/estuary/blob/master/LICENSE.md
  version: 0.0.0
servers:
- url: //api.estuary.tech/
security:
- bearerAuth: []
paths:
  /admin/autoretrieve/init:
    post:
      tags:
      - autoretrieve
      summary: Register autoretrieve server
      description: This endpoint registers a new autoretrieve server
      operationId: adminAutoretrieveInitPOST
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/autoretrieve_init_body'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Autoretrieve
  /admin/autoretrieve/list:
    get:
      tags:
      - autoretrieve
      summary: List autoretrieve servers
      description: This endpoint lists all registered autoretrieve servers
      operationId: adminAutoretrieveListGET
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Autoretrieve
  /admin/peering/peers:
    get:
      tags:
      - admin
      - peering
      - peers
      summary: List all Peering peers
      description: This endpoint can be used to list all peers on Peering Service
      operationId: adminPeeringPeersGET
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Admin
    post:
      tags:
      - admin
      - peering
      - peers
      summary: Add peers on Peering Service
      description: This endpoint can be used to add a Peer from the Peering Service
      operationId: adminPeeringPeersPOST
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Admin
    delete:
      tags:
      - admin
      - peering
      - peers
      summary: Remove peers on Peering Service
      description: This endpoint can be used to remove a Peer from the Peering Service
      operationId: adminPeeringPeersDELETE
      requestBody:
        description: Peer ids
        content:
          '*/*':
            schema:
              type: array
              items:
                type: boolean
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-codegen-request-body-name: peerIds
      x-swagger-router-controller: Admin
  /admin/peering/start:
    post:
      tags:
      - admin
      - peering
      - peers
      summary: Start Peering
      description: This endpoint can be used to start the Peering Service
      operationId: adminPeeringStartPOST
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Admin
  /admin/peering/status:
    get:
      tags:
      - admin
      - peering
      - peers
      summary: Check Peering Status
      description: This endpoint can be used to check the Peering status
      operationId: adminPeeringStatusGET
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Admin
  /admin/peering/stop:
    post:
      tags:
      - admin
      - peering
      - peers
      summary: Stop Peering
      description: This endpoint can be used to stop the Peering Service
      operationId: adminPeeringStopPOST
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Admin
  /admin/system/config:
    get:
      tags:
      - admin
      summary: Get systems(estuary/shuttle) config
      description: This endpoint is used to get system configs.
      operationId: adminSystemConfigGET
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Admin
  /admin/users:
    get:
      tags:
      - admin
      summary: Get all users
      description: This endpoint is used to get all users.
      operationId: adminUsersGET
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Admin
...
components:
  schemas:
...
    util.HttpError:
      type: object
      properties:
        code:
          type: integer
        details:
          type: string
        reason:
          type: string
    autoretrieve_init_body:
      required:
      - addresses
      - pubKey
      type: object
      properties:
        addresses:
          type: string
          description: Autoretrieve's comma-separated list of addresses
        pubKey:
          type: string
          description: Autoretrieve's public key
...
  securitySchemes:
    bearerAuth:
      type: apiKey
      name: Authorization
      in: header
...